requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==23.7.0
bleach==6.1.0
boto3==1.29.4
botocore==1.32.4
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.5
cssselect2==0.7.0
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==5.0.2
django-bootstrap-v5==1.0.11
django-bootstrap4==24.1
django-bootstrap5==23.4
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-rest-authtoken==2.1.4
django-storages==1.14.2
django-summernote==0.8.20.0
django-taggit==5.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
drf-yasg==1.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.0.7
mypy-extensions==1.0.0
mysqlclient==2.2.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
oscrypto==1.3.0
packaging==23.1
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2023.11.15
pipreqs==0.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.43
protobuf==4.25.1
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.18.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
redis==5.0.1
referencing==0.35.1
reportlab==4.0.9
requests==2.31.0
rpds-py==0.18.1
rsa==4.9
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.3.1
stack-data==0.6.3
storage==0.0.4.3
svglib==1.5.1
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.3
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.4
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.14
yarg==0.1.9
zope.interface==6.1
